Title: n개의 숫자 중 최소, Time: 16ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240805/n개의 숫자 중 최소/min-of-n-num.cpp
@@ -6,9 +6,9 @@ using namespace std;
 int main() {
 
     int arr[101];
-    int n;
-    int min = 9999;
-    int cnt = 0;
+    long int n;
+    long int min = 9999;
+    long int cnt = 0;
 
     cin >> n;
